+++ src/onegov.election_day/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Allows to upload UTF-16 wabsti files.
+  [msom]
+
 1.14.1 (2017-06-23)
 ~~~~~~~~~~~~~~~~~~~